backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
exceptiongroup==1.1.3
fastapi==0.103.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
lanarky==0.7.13
langchain==0.0.292
langsmith==0.0.37
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.25.2
openai==0.28.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pydantic==1.10.12
pydantic_core==2.6.3
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tenacity==8.2.3
tomli==2.0.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.23.2
yarl==1.9.2
